requirements-dev.txt
# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0
moto>=5.0.0  # AWS mocking

# Code Quality
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0

# Type Stubs
boto3-stubs[essential]>=1.34.0
pandas-stubs>=2.1.0
types-pyyaml>=6.0.0
types-requests>=2.31.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocstrings[python]>=0.24.0

# Debugging
ipython>=8.20.0
ipdb>=0.13.0
